user authentication

+++ routes/index.js
@@ -3,9 +3,20 @@ var router = express.Router();
 
 const myDB = require('../db/myDB.js');
 
+function auth(req, res) {
+  if (!req.session.username) {
+    res.status(401).send({ err: 'user not logged in' });
+    return false;
+  }
+  return true;
+}
+
 // Load data from db
 router.get('/getData', async (req, res) => {
   try {
+    if (!auth(req, res)) {
+      return;
+    }
     const dataArray = await myDB.getData();
     res.status(200).send({ data: dataArray });
   } catch (e) {
@@ -17,8 +28,10 @@ router.get('/getData', async (req, res) => {
 // save a car to user's collections
 router.post('/saveCar', async (req, res) => {
   try {
-    // TODO: get actual username from session
-    const username = 'xingyu711';
+    if (!auth(req, res)) {
+      return;
+    }
+    const username = req.session.username;
     const carId = req.body.car_id;
 
     await myDB.addToCollections(username, carId);
@@ -32,8 +45,10 @@ router.post('/saveCar', async (req, res) => {
 // unsave a car from user's collections
 router.post('/unsaveCar', async (req, res) => {
   try {
-    // TODO: get actual username from session
-    const username = 'xingyu711';
+    if (!auth(req, res)) {
+      return;
+    }
+    const username = req.session.username;
     const carId = req.body.car_id;
 
     await myDB.deleteFromCollections(username, carId);
@@ -47,8 +62,10 @@ router.post('/unsaveCar', async (req, res) => {
 // load data saved in user's collections
 router.get('/getCollections', async (req, res) => {
   try {
-    // TODO: get actual username from session
-    const username = 'xingyu711';
+    if (!auth(req, res)) {
+      return;
+    }
+    const username = req.session.username;
 
     const savedCars = await myDB.getUserCollections(username);
     res.status(200).send({ data: savedCars });
@@ -61,8 +78,11 @@ router.get('/getCollections', async (req, res) => {
 // user post info to db
 router.post('/postInfo', async (req, res) => {
   try {
-    // TODO: get actual username from session
-    const username = 'xingyu711';
+    if (!auth(req, res)) {
+      return;
+    }
+    // get username from session and add username to data object
+    const username = req.session.username;
 
     const dataObject = JSON.parse(JSON.stringify(req.body));
     dataObject['username'] = username;
@@ -80,8 +100,10 @@ router.post('/postInfo', async (req, res) => {
 // delete a post in user's home page
 router.post('/deletePost', async (req, res) => {
   try {
-    // TODO: get actual username from session
-    const username = 'xingyu711';
+    if (!auth(req, res)) {
+      return;
+    }
+    const username = req.session.username;
     const carId = req.body.car_id;
 
     await myDB.deleteFromPosts(username, carId);
@@ -94,12 +116,89 @@ router.post('/deletePost', async (req, res) => {
 
 router.get('/getPosts', async (req, res) => {
   try {
-    // TODO: get actual username from session
-    const username = 'xingyu711';
+    if (!auth(req, res)) {
+      return;
+    }
+    const username = req.session.username;
 
     const userPosts = await myDB.getUserPosts(username);
 
     res.status(200).send({ data: userPosts });
+  } catch (e) {
+    console.log('Error', e);
+    res.status(400).send({ err: e });
+  }
+});
+
+router.post('/userLogin', async (req, res) => {
+  try {
+    const username = req.body.username;
+    const password = req.body.password;
+    // ask db to validate this user
+    const msg = await myDB.validateUser(username, password);
+
+    // successfully loged in
+    if (msg === 'success') {
+      // save username to session
+      req.session.username = username;
+      res.sendStatus(200);
+    } else {
+      res.status(401).send({ login: msg });
+    }
+  } catch (e) {
+    console.log('Error', e);
+    res.status(400).send({ err: e });
+  }
+});
+
+router.post('/userRegister', async (req, res) => {
+  try {
+    const username = req.body.username;
+    const password = req.body.password;
+    const firstname = req.body.firstname;
+    const lastname = req.body.lastname;
+
+    const msg = await myDB.registerUser(
+      username,
+      password,
+      firstname,
+      lastname
+    );
+    if (msg === 'success') {
+      // save username to session
+      req.session.username = username;
+      res.sendStatus(200);
+    } else {
+      res.status(409).send({ register: msg });
+    }
+  } catch (e) {
+    console.log('Error', e);
+    res.status(400).send({ err: e });
+  }
+});
+
+router.get('/userLogout', async (req, res) => {
+  try {
+    if (!auth(req, res)) {
+      return;
+    }
+    delete req.session.username;
+    res.sendStatus(200);
+  } catch (e) {
+    console.log('Error', e);
+    res.status(400).send({ err: e });
+  }
+});
+
+router.get('/getPersonName', async (req, res) => {
+  try {
+    if (!auth(req, res)) {
+      return;
+    }
+    const username = req.session.username;
+    //ask db to find the user's display name
+    const displayName = await myDB.getUserDisplayName(username);
+    res.status(200).send({ displayName: displayName });
   } catch (e) {
     console.log('Error', e);
     res.status(400).send({ err: e });
